fix(SA2-156): wins は小数(チョップ)を許容する
ユーザー指摘対応。all-in の wins はスプリットポット(チョップ)で小数になり得る
ため、当初追加した .int() / 整数チェックは誤り。allInPayload と両フォーム
(use-all-in-form / use-all-in-editor)から整数制約を除去し、SA2-156 の本質で
ある wins <= trials の上限のみを維持する。trials は run-out 回数なので整数の
まま。

これにより、小数 wins を送る既存コンポーネントテスト
(all-in-bottom-sheet.test.tsx)を .int() が弾いていた CI 回帰も解消。テストは
「小数(チョップ)受け入れ」+「小数でも trials 超過は拒否」に更新。

+++ packages/db/src/constants/session-event-types.ts
@@ -99,17 +99,19 @@ export const chipsAddRemovePayload = z.object({
 		.refine((n) => n !== 0, { message: "amount must be non-zero" }),
 });
 
-// `wins` is the number of favorable outcomes across `trials` all-in run-outs,
-// so it must be a non-negative integer that never exceeds `trials`. Without the
-// `.int()` and the object-level `wins <= trials` refine, a payload like
-// `{ potSize: 1000, trials: 1, wins: 5 }` validated and let the EV math compute a
-// wins-share larger than the pot, corrupting `evCashOut` / `evDiff` (SA2-156).
+// `wins` is the number of favorable all-in run-outs across `trials`, counted as
+// a fraction when the pot is chopped (a split counts as a partial win). It is
+// therefore a non-negative number — NOT necessarily an integer — that never
+// exceeds `trials`. The object-level `wins <= trials` refine blocks the real bug:
+// a payload like `{ potSize: 1000, trials: 1, wins: 5 }` used to validate and let
+// the EV math compute a wins-share larger than the pot, corrupting
+// `evCashOut` / `evDiff` (SA2-156).
 export const allInPayload = z
 	.object({
 		potSize: z.number().min(0),
 		trials: z.number().int().min(1),
 		equity: z.number().min(0).max(100),
-		wins: z.number().int().min(0),
+		wins: z.number().min(0),
 	})
 	.refine((data) => data.wins <= data.trials, {
 		message: "wins must not exceed trials",
